Split up Supervisor into StreamSupervisor.
