+++ server/src/test/java/org/apache/druid/indexing/NoopSupervisorSpecTest.java
@@ -22,6 +22,7 @@ package org.apache.druid.indexing;
 import com.google.common.collect.ImmutableList;
 import org.apache.druid.indexing.overlord.ObjectMetadata;
 import org.apache.druid.indexing.overlord.supervisor.NoopSupervisorSpec;
+import org.apache.druid.indexing.overlord.supervisor.StreamSupervisor;
 import org.apache.druid.indexing.overlord.supervisor.Supervisor;
 import org.apache.druid.indexing.overlord.supervisor.SupervisorStateManager;
 import org.apache.druid.indexing.overlord.supervisor.autoscaler.LagStats;
@@ -40,16 +41,9 @@ public class NoopSupervisorSpecTest
     Exception e = null;
     try {
       NoopSupervisorSpec noopSupervisorSpec = new NoopSupervisorSpec(null, Collections.singletonList("datasource1"));
-      Supervisor supervisor = noopSupervisorSpec.createSupervisor();
+      StreamSupervisor supervisor = noopSupervisorSpec.createSupervisor();
       SupervisorTaskAutoScaler autoscaler = noopSupervisorSpec.createAutoscaler(supervisor);
       Assert.assertNull(autoscaler);
-      Callable<Integer> noop = new Callable<Integer>() {
-          @Override
-          public Integer call()
-          {
-            return -1;
-          }
-      };
 
       int count = supervisor.getActiveTaskGroupsCount();
       Assert.assertEquals(count, -1);
@@ -79,7 +73,7 @@ public class NoopSupervisorSpecTest
   public void testNoppSupervisorResetOffsetsDoNothing()
   {
     NoopSupervisorSpec expectedSpec = new NoopSupervisorSpec(null, null);
-    Supervisor noOpSupervisor = expectedSpec.createSupervisor();
+    StreamSupervisor noOpSupervisor = expectedSpec.createSupervisor();
     Assert.assertEquals(-1, noOpSupervisor.getActiveTaskGroupsCount());
     noOpSupervisor.resetOffsets(null);
     Assert.assertEquals(-1, noOpSupervisor.getActiveTaskGroupsCount());
@@ -95,7 +89,7 @@ public class NoopSupervisorSpecTest
   public void testNoppSupervisorStopTaskEarlyDoNothing()
   {
     NoopSupervisorSpec expectedSpec = new NoopSupervisorSpec(null, null);
-    Supervisor noOpSupervisor = expectedSpec.createSupervisor();
+    StreamSupervisor noOpSupervisor = expectedSpec.createSupervisor();
     Assert.assertThrows(UnsupportedOperationException.class,
         () -> noOpSupervisor.handoffTaskGroupsEarly(ImmutableList.of())
     );
